+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ChangeLog
 
+0.4.3:
+
+-added 'mod', 'pi' and 20 other unary functions (basically completing coverage of JavaScript Math library's unary functions)
+-added 'osc' as a synonym for the now deprecated 'sin'; as of version 0.5, 'osc' will be the oscillator, and 'sin' will be the sine function
+
 0.4.2:
 
 - webcam support officially added, via 'cam'